+++ app/Controllers/seccion.php
@@ -2,6 +2,7 @@
 
 use CodeIgniter\Controller;
 use App\Models\secciones;
+use App\Models\ofertas;
 
 class seccion extends BaseController
 {
@@ -80,14 +81,20 @@ class seccion extends BaseController
 	}
 	public function eliminar()
 	{
-		$id=$this->request->getPost('valor_id_seccion');   //varible que recive los valores de input valor_id_seccion
 		$valor=0;  
-		$secciones = new secciones();
-		$result = $secciones->where('id',$id)->delete();
-		if(!empty($result))
-		{
-			$valor=1;
+		$id=$this->request->getPost('valor_id_seccion');   //varible que recive los valores de input valor_id_seccion
+		$oferta = new ofertas();//creacion de instancia oferta para hacer la busqueda en la tabla		
+		$buscar = $oferta->where('Seccionid',$id)->find();//buscar si el Grupoid no existe en la tabla oferta
+
+		if($buscar==false)//si no existe en la tabla oferta
+		{		
+			$secciones = new secciones();
+			$result = $secciones->where('id',$id)->delete();
+			if(!empty($result))
+			{
+				$valor=1;
+			}
 		}
-		return  json_decode($valor);
+		return  json_encode($valor);
 	}
 }
